Review/V0.3.2/requirements.txt
# --- Audiopro Industrial Suite v0.3.1 Dependencies ---

# --- CORE UI & THREADING ---
PySide6==6.6.1

# --- SECURITY & VERIFICATION ---
python-magic==0.4.27; platform_system != "Windows"
python-magic-bin==0.4.14; platform_system == "Windows"
# Note: requires 'libmagic' installed on the OS (brew install libmagic / apt install libmagic1)

# --- MACHINE LEARNING & INTELLIGENCE ---
scikit-learn==1.3.2
joblib==1.3.2
requests==2.31.0
pandas==2.1.4

# --- CORE SIGNAL PROCESSING & SPECTRAL ANALYSIS ---
librosa==0.10.1
numpy==1.26.2
scipy==1.11.3
soundfile==0.12.1
mutagen==1.47.0

# METADATA & HEADER EXTRACTION
mutagen==1.47.0

# GPU & HARDWARE MONITORING
psutil==5.9.6
pynvml==11.5.0
